refactored run subcommand
